Extracting reusable logic from Query, QueryCacher and ResultSetDecorator out into the Datasource namespace
